Improve getField in Resource instead of creating getResourceField in kb
